refactor : AuthenticationService 추상화 & 네이밍 변경(redirectUri -> authorizationUri)
