+++ src/main/java/com/meoguri/linkocean/domain/user/service/AuthenticationService.java
@@ -1,35 +1,11 @@
 package com.meoguri.linkocean.domain.user.service;
 
-import org.springframework.stereotype.Service;
-
 import com.meoguri.linkocean.domain.user.entity.vo.Email;
 import com.meoguri.linkocean.domain.user.entity.vo.OAuthType;
 
-import lombok.RequiredArgsConstructor;
+public interface AuthenticationService {
 
-@RequiredArgsConstructor
-@Service
-public class AuthenticationService {
-
-	private final OAuthClient oAuthClient;
-
-	public String getRedirectUrl(final OAuthType oAuthType) {
-
-		if (oAuthType != OAuthType.GOOGLE) {
-			throw new IllegalArgumentException("구글 소셜 로그인만 지원되는 기능입니다.");
-		}
-
-		return oAuthClient.getRedirectUrl();
-	}
-
-	public Email authenticate(final OAuthType oAuthType, final String authorizationCode) {
-
-		//TODO 벤더사 추가 쉽도록 확장성 있게 리팩토링하기, 아직 구글만 지원함.
-		if (oAuthType != OAuthType.GOOGLE) {
-			throw new IllegalArgumentException("구글 소셜 로그인만 지원되는 기능입니다.");
-		}
-
-		final String accessToken = oAuthClient.getAccessToken(authorizationCode);
-		return oAuthClient.getUserEmail(accessToken);
-	}
+	String getAuthorizationUri(OAuthType oAuthType);
+	
+	Email authenticate(OAuthType oAuthType, String authorizationCode);
 }
