+++ src/main/java/com/meoguri/linkocean/domain/user/service/OAuthClient.java
@@ -6,7 +6,7 @@ public interface OAuthClient {
 
 	/* 프론트에서 가져갈 일 */
 	@Deprecated
-	String getRedirectUrl();
+	String getAuthorizationUri();
 
 	/* 인증 코드 이용해 access Token 얻어온다 */
 	String getAccessToken(String authorizationCode);
